+++ src/controllers/api/user.controller.js
@@ -1,6 +1,5 @@
 // src/controllers/api/user.controller.js
 
-import { getUserInfo } from "../../service/userProfileService.js";
 import fs from 'fs';
 import path from "path";
 import { fileURLToPath } from 'url';
@@ -12,15 +11,17 @@ const __filename = fileURLToPath(import.meta.url);
 const __dirname = dirname(__filename);
 
 // 個人頁面更改邏輯
-{/* 有Bug 假設A使用Google首次登入，個人相片會顯示google提供的沒問題，但當我在個人頁面修改按下儲存會導致avatar變成沒有東西avatorPath(邏輯需要修改) */}
+{/* 有Bug 假設A使用Google首次登入，個人相片會顯示google提供的沒問題，但當我在個人頁面修改按下儲存會導致avatar變成沒有東西avatorPath(邏輯需要修改) 
+    目前 前端傳到後端的時候，從network來看payload的avatar是空值*/}
 export const profileChange = async (req, res) => {
     const userId = req.params.id;
     const { name, email } = req.body;
     const avatarFile = req.file;
-    // console.log(avatarFile.filename);
+    console.log(avatarFile);
 
     try {
-        let avatarPath = null;
+        // 待avatar邏輯清晰再打開。
+        // let avatarPath = null;
 
         // 整合到multer.js
         const uploadDir = path.join(__dirname, '../../uploads'); // 確認uploads目錄的正確路徑
@@ -29,14 +30,14 @@ export const profileChange = async (req, res) => {
         }
 
         if (avatarFile) {
-            avatarPath = `/uploads/${avatarFile.filename}`;
+            // avatarPath = `/uploads/${avatarFile.filename}`;
         }
 
         const userToBeChanged = await User.findByPk(userId);
 
         if (userToBeChanged) {
             userToBeChanged.username = name;
-            userToBeChanged.avatar = avatarPath;
+            // userToBeChanged.avatar = avatarPath;
             await userToBeChanged.save();
 
             const userInfo = {
@@ -48,7 +49,7 @@ export const profileChange = async (req, res) => {
 
             return res.status(200).json({ status: "success", data: userInfo });
         } else {
-            return res.status(404).json({ status: "error", message: "Not Found" });
+            return res.status(404).json({ status: "error", message: "User Not Found" });
         }
     } catch (error) {
         console.error(error);
@@ -61,9 +62,22 @@ export const userInfo = async (req, res) => {
     const userId = req.userId;
 
     try {
-        const userInfo = await getUserInfo(userId);
-        res.status(200).json({ data: userInfo });
+        const userInfoBuffer = await User.findByPk(userId);
+
+        if (userInfoBuffer) {
+            const userData = {
+                id: userInfoBuffer.id,
+                name: userInfoBuffer.username,
+                email: userInfoBuffer.email,
+                avatar: userInfoBuffer.avatar
+            }
+
+            return res.status(200).json({ status: "success", data: userData });
+        } else {
+            return res.status(404).json({ status: "error", message: "User Not Found"})
+        }
     } catch (error) {
-        res.status(500).json({ message: 'An error occurred while fetching the user information', error });
+        console.error(error);
+        return res.status(500).json({ status: "error", message: "Internal server error" });
     }
 }
